moved resource package to the service level
